Separate problem text from JavaScript file
