+++ menu.json
@@ -1,0 +1,18 @@
+[
+  "Install npm",
+  "Dev Environment",
+  "Install A Module",
+  "Listing Dependencies",
+  "npm Test",
+  "Login",
+  "Package Niceties",
+  "Publish",
+  "Version",
+  "Publish Again",
+  "Dist Tag",
+  "Dist Tag Removal",
+  "Outdated",
+  "Update",
+  "Rm",
+  "Finale"
+]
