+++ problems/dev-environment/index.js
@@ -1,29 +1,12 @@
-// Old-sk00l multi-line strings
-exports.problem = function () { /*
-One of the most important things that npm does is install packages.
-
-However, in order to be a good workshop program, we don't want to litter
-files all over your computer, so before going any further, let's set up a
-development environment.
-
-First, make a new directory and `cd` into it.
-
-Then, run `npm init` to create a package.json file. You will be prompted
-to answer a number of questions. You can just press enter to accept the
-default for each question.
-
-Run `$ADVENTURE_COMMAND verify` once you're done.  All the other commands
-you run in this tutorial should be done in that folder.
-*/ }.toString().split('\n').slice(1, -1).join('\n')
-
-// exports.solution = function () {/*
-// mkdir test
-// cd test
-// */}.toString().split('\n').slice(1,-1).join('\n')
-
-var shop = require('../')
+var shop = require('../../')
 var fs = require('fs')
 var path = require('path')
+
+exports.init = function (workshopper) {
+  this.problem = {
+    file: path.join(__dirname, 'problem.{workshopper.lang}.txt')
+  }
+}
 
 exports.verify = function (args, cb) {
   var datadir = shop.datadir
