+++ problems/dist-tag-removal/problem.en.txt
@@ -1,0 +1,12 @@
+Now that you've added a dist-tag or two, let's clean things up.
+
+The only dist-tag you CAN'T ever remove is "latest".  That's because
+every package installs its "latest" tag by default, so that tag has
+some special semantics.
+
+You CAN point "latest" to a different version, or delete other tags.
+
+Let's delete all the tags that we can, and also point "latest" at
+something other than the most recent release.
+
+Run `npm help dist-tag` to learn more about the command.
