+++ problems/install-a-module/problem.en.txt
@@ -1,0 +1,14 @@
+The first thing that most people do with npm is install a dependency.
+
+Dependencies are fetched from the registry, and unpacked in the `node_modules`
+folder.
+
+To install a module, use the `npm install <modulename>` command.
+
+The registry that we're using for this tutorial is a tiny version of
+the one at https://registry.npmjs.org.  So you might find that it only has
+a small number of things.
+
+Let's start out by installing the "@linclark/pkg" module.
+
+Run `$ADVENTURE_COMMAND verify` once you're done.
