+++ problems/listing-dependencies/problem.en.txt
@@ -1,0 +1,10 @@
+npm isn't just for installing stuff.  It also shows you what you
+have installed (your dependencies).
+
+You can do this using the `npm ls` command.
+
+Run this command in your working dir.  If there are
+any problems npm will alert you by returning an "!ERR" message.
+
+If everything looks ok, then run `$ADVENTURE_COMMAND verify OK`, or
+`$ADVENTURE_COMMAND verify NOT OK` if npm does report a problem".
