+++ problems/publish-again/problem.en.txt
@@ -1,0 +1,9 @@
+Publishing something once is fine.  But healthy packages get
+published again and again with new and exciting bug fixes.
+
+You can't re-use the same version number again, because that's hella
+confusing for all the robots running the treadmills that power the npm
+registry.  But, now that we changed the version number in the last
+exercise, you can publish the package again.
+
+Go for it!  Then get your prize with `$ADVENTURE_COMMAND verify`
