+++ problems/start-a-project/problem.en.txt
@@ -1,0 +1,9 @@
+npm helps you build projects, but for npm to be able to do that, you need
+to tell npm a little bit about your project. You can tell npm about your
+project in a file called package.json.
+
+Run `npm init --scope=<username>`, and replace <username> with the user
+you created in the last lesson. This will create a package.json file.
+(For extra credit, set the project up with a git repo as well.)
+
+Run `$ADVENTURE_COMMAND verify` once you're done.
